Address failing sync in shared memory access
On some systems relying on a shared memory file object and its
memory mapped representation seems to not be fully syncronous.
In such a case, writing to the file descriptor and then immediately
read from the memory mapped buffer is a hit and miss.

This patch eliminates the memomry mapped buffer and replaces it with
a separate heap allocated buffer that just reads from the file
descriptor.