Fix parsing of shorter lines after long ones
The file descriptor has to be reinitialized for each line because
simply setting the pointer to the beginning keeps the file size
to the biggest one set up to the moment. This is not what we want
as we should know the total number of charaters the template
parsing processed.